# docs/api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Venora API",
    "version": "2026-07-14",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    },
    "summary": "Code-backed HTTP API for Venora",
    "description": "Next.js Route Handlers and Supabase Edge Functions confirmed at commit 15e6173b6f695f6d9f5f4a29517badf0b3a3a016. Server Actions and PostgreSQL RPCs are documented separately and are not represented as fake REST routes."
  },
  "servers": [
    {
      "url": "{appUrl}",
      "description": "Venora Next.js application",
      "variables": {
        "appUrl": {
          "default": "http://localhost:3000",
          "description": "NEXT_PUBLIC_APP_URL"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/payouts/reconcile": {
      "post": {
        "summary": "Reconcile stuck withdrawals",
        "description": "Scheduled sweep. Resolves withdrawals left in `processing` against PayMongo's authoritative transfer status. Withdrawals with no transfer id are checked against the provider before any decision, because an ambiguous create (timeout or 5xx) must never release funds on assumption. Authenticated with the PAYOUT_RECONCILE_SECRET shared secret compared in constant time.",
        "operationId": "postApiPayoutsReconcile",
        "tags": ["Payments"],
        "x-venora-surface": "next-route-handler",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["received", "result"],
                  "properties": {
                    "received": {
                      "const": true
                    },
                    "result": {
                      "type": "string",
                      "enum": ["processed", "duplicate", "skipped"]
                    }
                  }
                },
                "example": {
                  "received": true,
                  "result": "processed"
                }
              }
            }
          },
          "401": {
            "description": "Invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Invalid signature"
                }
              }
            }
          },
          "500": {
            "description": "Processing failed; provider should retry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Processing failed"
                }
              }
            }
          },
          "503": {
            "description": "Gateway not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Provider not configured"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase JWT",
        "description": "Supabase access token. Never use a service-role token in browser code."
      },
      "supabaseSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-<project-ref>-auth-token",
        "description": "Supabase SSR session cookie; the actual value can be chunked."
      },
      "paymongoSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Paymongo-Signature",
        "description": "Timestamped PayMongo HMAC over the raw body."
      }
    },
...
  }
}